[FIX] prevent traceback when using the SIGIO signal
Before this commit, using the SIGIO posix signal to get server stats
would generate a traceback and was preventing the logging of server
 stats.
This was caused by a typo in the name of one property access during
destructuring.

This traceback is not critical and would not lead to server failure.