[IMP] Add per-channel encryption feature
This commit adds per-channel encryption to prevent users from signing
JWTs for channels that they did not create.

This features works by allowing an additional `key` claim in the JWT
used to request a room. When this claim is set, the JWTs used to
authenticate users are expected to be signed by this new key and not
the "global" SFU key (`AUTH_KEY` global variable).

This change is useful for cases like Odoo.SH where multiple parties
can use the same SFU with the same credentials.

task-3861455