[FIX] prevent potential race condition during websocket init
Before this commit, the connection of the rtc session on the server,
which was dependent on the speed of asynchronous mediasoup calls
(`router.createWebRtcTransport`) could race the initialization of the
websocket on the client.

This commit fixes this issue by signaling to the client when the
websocket is authenticated as soon as possible, before starting
the RTC connection process.